+++ package.json
@@ -8,7 +8,8 @@
     "start": "next start",
     "lint": "next lint",
     "type:check": "tsc --noEmit",
-    "analyze": "cross-env ANALYZE=true next build"
+    "analyze": "cross-env ANALYZE=true next build",
+    "optimization": "yarn build && yarn start"
   },
   "dependencies": {
     "@tanstack/react-query": "^4.29.5",
